De-unified Command struct for build
