+++ src/lib.rs
@@ -15,9 +15,6 @@ pub const MAX_COMMANDS_PER_STATE: usize = 3;
 
 use serde::{Deserialize, Serialize};
 
-#[cfg(feature = "executing")]
-use core::sync::atomic::AtomicBool;
-
 #[derive(Debug, Serialize, Deserialize, Copy, Clone, PartialEq)]
 pub struct Seconds(pub f32);
 
@@ -64,29 +61,4 @@ pub enum CommandObject {
     Pyro3(bool),
     Beacon(bool),
     DataRate(u16),
-}
-
-/// An action that takes place at a specific time after the state containing this is entered
-#[derive(Debug, Serialize, Deserialize, Clone, PartialEq)]
-pub struct Command {
-    /// The object that this command will act upon
-    pub object: crate::CommandObject,
-
-    /// How long after the state activates to execute this command
-    pub delay: crate::Seconds,
-
-    /// If this command has already executed
-    #[cfg(feature = "executing")]
-    pub was_executed: AtomicBool,
-}
-
-impl Command {
-    pub fn new(object: crate::CommandObject, delay: crate::Seconds) -> Self {
-        Self {
-            object,
-            delay,
-            #[cfg(feature = "executing")]
-            was_executed: AtomicBool::new(false),
-        }
-    }
 }
